feat(secrets): migrate concrete stores + agent_provider to dasclaw_runtime
Moves the four remaining ironclaw-internal secrets modules into the
host-agnostic crate so future hosts (codex, claw-code, headless) can
reuse the encrypted multi-user secrets stack without re-implementing
~1300 lines:

  - secrets::store_in_memory::InMemorySecretsStore  (~135 LOC)
  - secrets::store_postgres::PostgresSecretsStore   (~250 LOC, feature postgres)
  - secrets::store_libsql::LibSqlSecretsStore       (~330 LOC, feature libsql)
  - secrets::agent_provider::AgentSecrets adapter   (~100 LOC)
    implementing dasclaw_core::SecretProvider over any SecretsStore

Adds dasclaw_runtime[postgres] / dasclaw_runtime[libsql] features so
ironclaw's existing postgres/libsql feature flags forward through.
The ironclaw secrets/store.rs and secrets/agent_provider.rs collapse
to thin re-export shims (so ~20 existing call sites keep working
unchanged); the existing #[cfg(test)] mod tests in store.rs stays in
the shim because it depends on ironclaw's testing::credentials.

Tests:
  cargo nextest run -p dasclaw_runtime --features postgres,libsql
    -> 80 passed, 0 skipped  (was 71)
  cargo nextest run -p dasclaw --lib secrets
    -> 54 passed, 0 failed   (all pre-existing ironclaw secrets tests
                              continue to pass through the shim)
  fmt / clippy --all-features / check_no_panics / check_no_new_ironclaw_literal: clean

3-layer review (subagent): 0 BLOCKER / 5 NOTE / 5 NIT.
All NOTES are verbatim-port artifacts (e.g. is_accessible doesn't
pre-check expiry — final get() gate still Fail-Safe; libsql usage_count
unwrap_or(0) — telemetry only) and are explicitly out of PR 4c scope.

Closes #641 (F3.2 phase 2 PR 4c of 6).
